Moved test execution code into engine
