+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io.kotest.engine/test/TestCaseExecutor.kt
@@ -1,22 +1,21 @@
-package io.kotest.core.internal
+package io.kotest.engine.test
 
 import io.kotest.core.TimeoutExecutionContext
 import io.kotest.core.extensions.TestCaseExtension
 import io.kotest.core.extensions.resolvedTestCaseExtensions
+import io.kotest.core.internal.isEnabled
 import io.kotest.core.spec.invokeAfterInvocation
 import io.kotest.core.spec.invokeAllAfterTestCallbacks
 import io.kotest.core.spec.invokeAllBeforeTestCallbacks
 import io.kotest.core.spec.invokeBeforeInvocation
 import io.kotest.core.test.NestedTest
 import io.kotest.core.test.TestCase
-import io.kotest.core.test.TestCaseExecutionListener
 import io.kotest.core.test.TestContext
 import io.kotest.core.test.TestResult
 import io.kotest.core.test.TestStatus
 import io.kotest.core.test.TestType
 import io.kotest.core.test.resolvedInvocationTimeout
 import io.kotest.core.test.resolvedTimeout
-import io.kotest.core.test.withCoroutineContext
 import io.kotest.fp.Try
 import io.kotest.mpp.log
 import io.kotest.mpp.replay
