+++ kotest-framework/kotest-framework-engine/src/jvmMain/kotlin/io/kotest/engine/listener/TestCaseListenerToTestEngineListenerAdapter.kt
@@ -1,6 +1,6 @@
 package io.kotest.engine.listener
 
-import io.kotest.core.test.TestCaseExecutionListener
+import io.kotest.engine.test.TestCaseExecutionListener
 import io.kotest.core.test.TestCase
 import io.kotest.core.test.TestResult
 
@@ -8,7 +8,7 @@ import io.kotest.core.test.TestResult
  * Converts events fired to a [TestCaseExecutionListener] into events fired to a [TestEngineListener]
  */
 class TestCaseListenerToTestEngineListenerAdapter(val listener: TestEngineListener) :
-    TestCaseExecutionListener {
+   TestCaseExecutionListener {
 
    override fun testFinished(testCase: TestCase, result: TestResult) {
       listener.testFinished(testCase, result)
